# api/docs/swagger.yaml
definitions:
...
  story.Author:
    properties:
      full_name:
        type: string
      user_id:
        type: string
      username:
        type: string
    type: object
  story.CommentStoryReq:
    properties:
      author_id:
        type: string
      content:
        type: string
      story_id:
        type: string
    type: object
  story.CommentStoryRes:
    properties:
      author_id:
        type: string
      content:
        type: string
      created_at:
        type: string
      id:
        type: string
      story_id:
        type: string
    type: object
  story.Comments:
    properties:
      author:
        $ref: '#/definitions/story.Author'
      content:
        type: string
      created_at:
        type: string
      id:
        type: string
    type: object
  story.CreateStoriesRequest:
    properties:
      content:
        type: string
      location:
        type: string
      tags:
        items:
          type: string
        type: array
      title:
        type: string
      user_id:
        type: string
    type: object
  story.CreateStoriesResponse:
    properties:
      author_id:
        type: string
      content:
        type: string
      created_at:
        type: string
      id:
        type: string
      location:
        type: string
      tags:
        items:
          type: string
        type: array
      title:
        type: string
    type: object
  story.GetAllStoriesRes:
    properties:
      limit:
        type: integer
      offset:
        type: integer
      stories:
        items:
          $ref: '#/definitions/story.Stories'
        type: array
      total:
        type: integer
    type: object
  story.GetCommentsOfStoryRes:
    properties:
      comments:
        items:
          $ref: '#/definitions/story.Comments'
        type: array
      limit:
        type: integer
      offset:
        type: integer
      total:
        type: integer
    type: object
  story.GetStoryRes:
    properties:
      author:
        $ref: '#/definitions/story.Author'
      comments_count:
        type: integer
      content:
        type: string
      created_at:
        type: string
      id:
        type: string
      likes_count:
        type: integer
      location:
        type: string
      tags:
        items:
          type: string
        type: array
      title:
        type: string
      updated_at:
        type: string
    type: object
  story.LikeRes:
    properties:
      liked_at:
        type: string
      story_id:
        type: string
      user_id:
        type: string
    type: object
  story.Stories:
    properties:
      author:
        $ref: '#/definitions/story.Author'
      comments_count:
        type: integer
      likes_count:
        type: integer
      location:
        type: string
      story_id:
        type: string
      title:
        type: string
    type: object
  story.UpdateStoriesReq:
    properties:
      content:
        type: string
      id:
        type: string
      title:
        type: string
    type: object
  story.UpdateStoriesRes:
    properties:
      author_id:
        type: string
      content:
        type: string
      id:
        type: string
      location:
        type: string
      tags:
        items:
          type: string
        type: array
      title:
        type: string
      updated_at:
        type: string
    type: object
info:
  contact: {}
paths:
...
  /api/v1/stories:
    get:
      description: get all stories
      parameters:
      - description: Number of stories to fetch
        in: query
        name: limit
        type: string
      - description: Number of stories to omit
        in: query
        name: offset
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/story.GetAllStoriesRes'
        "400":
          description: Invalid data
          schema:
            type: string
        "500":
          description: Server error
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: get all story
      tags:
      - stories
    post:
      description: create new story
      parameters:
      - description: info
        in: body
        name: info
        required: true
        schema:
          $ref: '#/definitions/story.CreateStoriesRequest'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/story.CreateStoriesResponse'
        "400":
          description: Invalid data
          schema:
            type: string
        "500":
          description: Server error
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: create story
      tags:
      - stories
  /api/v1/stories/{story_id}:
    delete:
      description: Delete new story
      parameters:
      - description: story_id
        in: path
        name: story_id
        required: true
        type: string
      responses:
        "200":
          description: succesfully deleted
          schema:
            type: string
        "400":
          description: Invalid data
          schema:
            type: string
        "500":
          description: Server error
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Delete story
      tags:
      - stories
    get:
      description: Get story by id
      parameters:
      - description: story_id
        in: path
        name: story_id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/story.GetStoryRes'
        "400":
          description: Invalid data
          schema:
            type: string
        "500":
          description: Server error
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Get story
      tags:
      - stories
    put:
      description: Update new story
      parameters:
      - description: story_id
        in: path
        name: story_id
        required: true
        type: string
      - description: info
        in: body
        name: info
        required: true
        schema:
          $ref: '#/definitions/story.UpdateStoriesReq'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/story.UpdateStoriesRes'
        "400":
          description: Invalid data
          schema:
            type: string
        "500":
          description: Server error
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Update story
      tags:
      - stories
  /api/v1/stories/{story_id}/comments:
    get:
      description: get comment of story
      parameters:
      - description: story_id
        in: path
        name: story_id
        required: true
        type: string
      - description: Number of stories to fetch
        in: query
        name: limit
        type: string
      - description: Number of stories to omit
        in: query
        name: offset
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/story.GetCommentsOfStoryRes'
        "400":
          description: Invalid data
          schema:
            type: string
        "500":
          description: Server error
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: comment of story
      tags:
      - stories
    post:
      description: comment to story
      parameters:
      - description: story_id
        in: path
        name: story_id
        required: true
        type: string
      - description: story_id
        in: body
        name: info
        required: true
        schema:
          $ref: '#/definitions/story.CommentStoryReq'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/story.CommentStoryRes'
        "400":
          description: Invalid data
          schema:
            type: string
        "500":
          description: Server error
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: comment story
      tags:
      - stories
  /api/v1/stories/{story_id}/like:
    post:
      description: comment to story
      parameters:
      - description: story_id
        in: path
        name: story_id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/story.LikeRes'
        "400":
          description: Invalid data
          schema:
            type: string
        "500":
          description: Server error
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: comment story
      tags:
      - stories
...
securityDefinitions:
  ApiKeyAuth:
    description: API Gateway
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"